Remove pydoctor.epydoc.apidoc module
The only functionality from this module that was still in use is the
UNKNOWN object, so I moved that to where it is used (the pyval_repr
module).
